# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OmniSupport API
  description: API documentation for the OmniSupport customer support backend.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/v1/tickets/{id}:
    get:
      tags:
        - Tickets
      summary: Get ticket details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          $ref: '#/components/responses/TicketResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Tickets
      summary: Update ticket fields
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        '200':
          $ref: '#/components/responses/TicketResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User identifier
...
    TicketId:
      $ref: '#/components/parameters/UserId'
...
  schemas:
...
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 10
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        category:
          type: string
          enum: [BILLING, TECHNICAL, ACCOUNT, REFUND, SHIPPING, GENERAL, OTHER]
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlationId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
  responses:
    UnauthorizedError:
      description: Authorization required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...